Add better error handling; Send sensor data type
For now, BadSerialMessageError is logged to command line, but ignored.
There are garbage messages that get through after stopping the serial
loop, these now do not cause the app to crash.

The datatype of the sensor reading is now sent in the format message.
This allows attempting to parse it, and catching any errors, before
inserting into the database.